+++ src/constants/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    HIRO_BASE_URL : "https://api.mainnet.hiro.so/extended/v1/",
+    HIRO_BASE_TEST_URL : "https://api.testnet.hiro.so/extended/v1/",
+}