+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Added getAccounts() method to fetch list of generated accounts
 - Added importWallet() to import account using privateKey
 - Added test suite to test wallet
+- Added get balance method to fetch STX balance of an account and test cases